Seperate PhpCSTaskFormatterElement class out to separate file.
