+++ src/Phing/Task/Optional/PhpCSTask.php
@@ -215,66 +215,3 @@ class PhpCSTask extends Task
     }
 }
 
-class PhpCSTaskFormatterElement extends \Phing\Type\DataType
-{
-    /**
-     * Type of output to generate
-     * @var string
-     */
-    protected $type = "";
-
-    /**
-     * Output to file?
-     * @var bool
-     */
-    protected $useFile = true;
-
-    /**
-     * Output file.
-     * @var string
-     */
-    protected $outfile = "";
-
-    /**
-     * Validate config.
-     */
-    public function parsingComplete()
-    {
-        if (empty($this->type)) {
-            throw new BuildException("Format missing required 'type' attribute.");
-        }
-        if ($this->useFile && empty($this->outfile)) {
-            throw new BuildException("Format requires 'outfile' attribute when 'useFile' is true.");
-        }
-    }
-
-    public function setType($type)
-    {
-        $this->type = $type;
-    }
-
-    public function getType()
-    {
-        return $this->type;
-    }
-
-    public function setUseFile($useFile)
-    {
-        $this->useFile = $useFile;
-    }
-
-    public function getUseFile()
-    {
-        return $this->useFile;
-    }
-
-    public function setOutfile($outfile)
-    {
-        $this->outfile = $outfile;
-    }
-
-    public function getOutfile()
-    {
-        return $this->outfile;
-    }
-}
